requirements.txt
torch-model-archiver==0.6.0
torch-workflow-archiver==0.2.4
torchserve==0.2.0
nvgpu==0.9.0
fastapi[all]
uvicorn[standard]
python-dotenv
requests
nltk
numpy
opencv-python==4.7.0.68
